backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.3
sqlalchemy==2.0.31
psycopg[binary]==3.2.1
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
